+++ museumpy/client.py
@@ -36,12 +36,12 @@ SEARCH_TEMPLATE = """<?xml version="1.0" encoding="UTF-8"?>
 
 
 class MuseumPlusClient(object):
-    def __init__(self, base_url=None, map_function=None, requests_kwargs=None):
+    def __init__(self, base_url=None, map_function=None, session=None):
         self.session = requests.Session()
         self.base_url = base_url
         self.xmlparser = xmlparse.XMLParser()
         self.map_function = map_function
-        self.requests_kwargs = requests_kwargs or {}
+        self.session = session or requests.Session()
 
     def fulltext_search(self, query, module='Object', limit=100, offset=0):
         url = f"{self.base_url}/ria-ws/application/module/{module}/search"
@@ -49,7 +49,7 @@ class MuseumPlusClient(object):
             'module_name': module,
             'query': query,
         }
-        data_loader = DataPoster(url, params, FULLTEXT_TEMPLATE, self.requests_kwargs)
+        data_loader = DataPoster(url, self.session, params, FULLTEXT_TEMPLATE)
         return response.SearchResponse(data_loader, limit, offset, self.map_function)
 
     def search(self, field, value, module='Object', limit=100, offset=0):
@@ -59,12 +59,12 @@ class MuseumPlusClient(object):
             'field': field,
             'value': value,
             }
-        data_loader = DataPoster(url, params, SEARCH_TEMPLATE, self.requests_kwargs)
+        data_loader = DataPoster(url, self.session, params, SEARCH_TEMPLATE)
         return response.SearchResponse(data_loader, limit, offset, self.map_function)
 
     def module_item(self, id, module='Object'):
         url = f"{self.base_url}/ria-ws/application/module/{module}/{id}"
-        data_loader = DataLoader(url, self.requests_kwargs)
+        data_loader = DataLoader(url, self.session)
         resp = response.SearchResponse(data_loader, map_function=self.map_function)
         if resp.count == 1:
             return resp[0]
@@ -72,18 +72,17 @@ class MuseumPlusClient(object):
 
     def download_attachment(self, id, module='Object', dir='.'):
         url = f"{self.base_url}/ria-ws/application/module/{module}/{id}/attachment"
-        data_loader = DataLoader(url, self.requests_kwargs)
+        data_loader = DataLoader(url, self.session)
         return data_loader.download_file(url, dir)
 
 
 class DataPoster(object):
-    def __init__(self, url, params=None, template=None, requests_kwargs=None):
-        self.session = requests.Session()
+    def __init__(self, url, session, params=None, template=None):
+        self.session = session
         self.url = url
         self.params = params
         self.template = template
         self.xmlparser = xmlparse.XMLParser()
-        self.requests_kwargs = requests_kwargs or {}
 
     def load(self, **kwargs):
         self.params.update(kwargs)
@@ -95,13 +94,12 @@ class DataPoster(object):
         res = self._post_content(url, xml, headers)
         return self.xmlparser.parse(res.content)
 
-    def _post_content(self, url, data, headers):
+    def _post_content(self, url, data, headers={}):
+        self.session.headers.update(headers)
         try:
             res = self.session.post(
                 url,
-                data=data,
-                headers=headers,
-                **self.requests_kwargs
+                data=data
             )
             res.raise_for_status()
         except requests.exceptions.HTTPError as e:
@@ -113,11 +111,10 @@ class DataPoster(object):
 
 
 class DataLoader(object):
-    def __init__(self, url, requests_kwargs=None):
-        self.session = requests.Session()
+    def __init__(self, url, session):
+        self.session = session
         self.url = url
         self.xmlparser = xmlparse.XMLParser()
-        self.requests_kwargs = requests_kwargs or {}
 
     def load(self, **kwargs):
         xml = self._get_xml(self.url)
@@ -140,12 +137,9 @@ class DataLoader(object):
         return self.xmlparser.parse(res.content)
 
     def _get_content(self, url, headers={}):
+        self.session.headers.update(headers)
         try:
-            res = self.session.get(
-                url,
-                headers=headers,
-                **self.requests_kwargs
-            )
+            res = self.session.get(url)
             res.raise_for_status()
         except requests.exceptions.HTTPError as e:
             raise errors.MuseumPlusError("HTTP error: %s" % e)